+++ Src/GraphicLibs/SFML/SfmlLibrary.hpp
@@ -13,7 +13,7 @@
 #include <SFML/System.hpp>
 #include <SFML/Window.hpp>
 
-/// \brief SFML library using float as template value from IGraph
+/// \brief SFML library
 class SfmlLibrary : public IGraph {
     public:
         /// \brief Default Constructor
@@ -25,14 +25,14 @@ class SfmlLibrary : public IGraph {
         bool clearWindow() override;
         bool displayWindow() override;
 
-        bool drawCircle(std::pair<float, float> pos, float radius,
+        bool drawCircle(std::pair<int, int> pos, int radius,
                         Color color) override;
-        bool drawRect(std::pair<float, float> pos, float width, float height,
+        bool drawRect(std::pair<int, int> pos, int width, int height,
                       Color color) override;
-        bool drawText(std::pair<float, float> pos,
+        bool drawText(std::pair<int, int> pos,
                       const std::string &content) override;
         bool drawEntity(IEntity &entity,
-                        std::pair<float, float> pos) override;
+                        std::pair<int, int> pos) override;
 
     private:
         /// \brief window member used to display every entities
